Switching to GNU Affero GPL v3.
-------------------------------

For a long time now I've been interested in open source licensing models.
From personal research on the subject, I've come to find two similar but
distinct models for licensing open source software:

  * The Apache License 2.0 model
  * The AGPL 3.0 model

Leaving patents, attribution, linking, and other nuances out of the
discussion, the main question for an open source contributor is:

  Do I want to allow others complete freedom on how they use my
  contributions, by adopting the Apache 2.0 License, or do I want
  my contributions to be used by others as long as they also
  contribute in kind, by adopting the GNU General Public License?

With the former, you open the door to allowing commercially motivated
contributors to use your software in hopes to drive adoption and
build a community, while opening yourself up to the possibility of
forking and closed source modifications and competitors.

With the latter, you accept that you may drive certain contributors
away from your project by requiring that they contribute their
modifications back to the OSS community, as they may see this as a
restriction on their free will to use the code in any way they see
fit.

As a software developer myself, I know that writing software requires
my time, which is not without some cost.  The pact with the OSS
community is that you may take advantage of the time I've spent
working on a piece of software, as long as I get to take advantage
of the time you spent on another piece of software.  It's almost
like a barter system, except we're bartering with time in a digital
medium.

Where this ideal system becomes complicated is when companies try
to monetize OSS software.  The OSS model is meant to facilitate open
sharing of software, which is in conflict with writing software for
monetary gain. Some companies avoid this by switching their business
model to a service-based model.  They don't sell the software, they
sell their expertise around the software.  Others try to differenti-
ate between an OSS version of their software and a commercial
version with more features.

Recently, this kind of competition between Elasticsearch and Amazon
has brought to light this very issue (see: https://onezero.medium.com/open-source-betrayed-industry-leaders-accuse-amazon-of-playing-a-rigged-game-with-aws-67177bc748b7?gi=356d74c0b36e)

As far as this project is concerned:

* I don't currently have aspirations to monetize this software.
* I don't care if this software gains wide adoption.
* I would love to be compensated for my time spent on this
  project, but I don't care if I don't.
* I would like others that care about this software to contribute
  back to it.
* I would like this software to remain open source in the future.

Given the above goals, I feel that I should change the license from
Apache 2.0 to GNU Affero General Public License 3.0.